# src/code/Umbraco.Community.PagespeedOptimizer.BackOffice.Client/src/api/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: `${string}://swagger.json` | (string & {});
};

export type CreateMediaExceptionRequestModel = {
    mediaKey: string;
    quality: number;
    forceWebp: boolean;
};

export type DefaultValuesResponseModel = {
    defaultImageQuality: number;
    enabled: boolean;
    forceWebP: boolean;
};

export type EventMessageTypeModel = 'Default' | 'Info' | 'Error' | 'Success' | 'Warning';

export type MediaExceptionResponseModel = {
    id: string;
    mediaKey: string;
    quality: number;
    forceWebp: boolean;
};

export type NotificationHeaderModel = {
    message: string;
    category: string;
    type: EventMessageTypeModel;
};

export type UpdateMediaExceptionRequestModel = {
    quality: number;
    forceWebp: boolean;
};

export type CreateMediaExceptionData = {
    body?: CreateMediaExceptionRequestModel;
    path?: never;
    query?: never;
    url: '/umbraco/pagespeed-optimizer/v1/media-exception';
};

export type CreateMediaExceptionErrors = {
    /**
     * The resource is protected and requires an authentication token
     */
    401: unknown;
    /**
     * The authenticated user does not have access to this resource
     */
    403: unknown;
    /**
     * Conflict
     */
    409: unknown;
};

export type CreateMediaExceptionResponses = {
    /**
     * Created
     */
    201: MediaExceptionResponseModel;
};

export type CreateMediaExceptionResponse = CreateMediaExceptionResponses[keyof CreateMediaExceptionResponses];

export type DeleteMediaExceptionData = {
    body?: never;
    path: {
        id: string;
    };
    query?: never;
    url: '/umbraco/pagespeed-optimizer/v1/media-exception/{id}';
};

export type DeleteMediaExceptionErrors = {
    /**
     * The resource is protected and requires an authentication token
     */
    401: unknown;
    /**
     * The authenticated user does not have access to this resource
     */
    403: unknown;
    /**
     * Not Found
     */
    404: unknown;
};

export type DeleteMediaExceptionResponses = {
    /**
     * OK
     */
    200: unknown;
};

export type UpdateMediaExceptionData = {
    body?: UpdateMediaExceptionRequestModel;
    path: {
        id: string;
    };
    query?: never;
    url: '/umbraco/pagespeed-optimizer/v1/media-exception/{id}';
};

export type UpdateMediaExceptionErrors = {
    /**
     * The resource is protected and requires an authentication token
     */
    401: unknown;
    /**
     * The authenticated user does not have access to this resource
     */
    403: unknown;
    /**
     * Not Found
     */
    404: unknown;
};

export type UpdateMediaExceptionResponses = {
    /**
     * OK
     */
    200: MediaExceptionResponseModel;
};

export type UpdateMediaExceptionResponse = UpdateMediaExceptionResponses[keyof UpdateMediaExceptionResponses];

export type GetByMediaKeyData = {
    body?: never;
    path: {
        mediaKey: string;
    };
    query?: never;
    url: '/umbraco/pagespeed-optimizer/v1/media-exception/by-media-key/{mediaKey}';
};

export type GetByMediaKeyErrors = {
    /**
     * The resource is protected and requires an authentication token
     */
    401: unknown;
    /**
     * The authenticated user does not have access to this resource
     */
    403: unknown;
};

export type GetByMediaKeyResponses = {
    /**
     * OK
     */
    200: MediaExceptionResponseModel | null;
};

export type GetByMediaKeyResponse = GetByMediaKeyResponses[keyof GetByMediaKeyResponses];

export type GetDefaultValuesData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/umbraco/pagespeed-optimizer/v1/media-exception/default-values';
};

export type GetDefaultValuesErrors = {
    /**
     * The resource is protected and requires an authentication token
     */
    401: unknown;
    /**
     * The authenticated user does not have access to this resource
     */
    403: unknown;
};

export type GetDefaultValuesResponses = {
    /**
     * OK
     */
    200: DefaultValuesResponseModel;
};

export type GetDefaultValuesResponse = GetDefaultValuesResponses[keyof GetDefaultValuesResponses];