# src/code/Umbraco.Community.PagespeedOptimizer.BackOffice.Client/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "PageSpeed Optimizer Management API",
    "version": "1.0"
  },
  "paths": {
    "/umbraco/pagespeed-optimizer/v1/media-exception": {
      "post": {
        "tags": [
          "Page Speed Optimizer"
        ],
        "operationId": "CreateMediaException",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/CreateMediaExceptionRequestModel"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/CreateMediaExceptionRequestModel"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/CreateMediaExceptionRequestModel"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "headers": {
              "Umb-Notifications": {
                "description": "The list of notifications produced during the request.",
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationHeaderModel"
                  },
                  "nullable": true
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/MediaExceptionResponseModel"
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "headers": {
              "Umb-Notifications": {
                "description": "The list of notifications produced during the request.",
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationHeaderModel"
                  },
                  "nullable": true
                }
              }
            }
          },
          "401": {
            "description": "The resource is protected and requires an authentication token"
          },
          "403": {
            "description": "The authenticated user does not have access to this resource",
            "headers": {
              "Umb-Notifications": {
                "description": "The list of notifications produced during the request.",
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationHeaderModel"
                  },
                  "nullable": true
                }
              }
            }
          }
        },
        "security": [
          {
            "Backoffice-User": [ ]
          }
        ]
      }
    },
    "/umbraco/pagespeed-optimizer/v1/media-exception/{id}": {
      "delete": {
        "tags": [
          "Page Speed Optimizer"
        ],
        "operationId": "DeleteMediaException",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Umb-Notifications": {
                "description": "The list of notifications produced during the request.",
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationHeaderModel"
                  },
                  "nullable": true
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "headers": {
              "Umb-Notifications": {
                "description": "The list of notifications produced during the request.",
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationHeaderModel"
                  },
                  "nullable": true
                }
              }
            }
          },
          "401": {
            "description": "The resource is protected and requires an authentication token"
          },
          "403": {
            "description": "The authenticated user does not have access to this resource",
            "headers": {
              "Umb-Notifications": {
                "description": "The list of notifications produced during the request.",
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationHeaderModel"
                  },
                  "nullable": true
                }
              }
            }
          }
        },
        "security": [
          {
            "Backoffice-User": [ ]
          }
        ]
      },
      "put": {
        "tags": [
          "Page Speed Optimizer"
        ],
        "operationId": "UpdateMediaException",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/UpdateMediaExceptionRequestModel"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/UpdateMediaExceptionRequestModel"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/UpdateMediaExceptionRequestModel"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Umb-Notifications": {
                "description": "The list of notifications produced during the request.",
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationHeaderModel"
                  },
                  "nullable": true
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/MediaExceptionResponseModel"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "headers": {
              "Umb-Notifications": {
                "description": "The list of notifications produced during the request.",
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationHeaderModel"
                  },
                  "nullable": true
                }
              }
            }
          },
          "401": {
            "description": "The resource is protected and requires an authentication token"
          },
          "403": {
            "description": "The authenticated user does not have access to this resource",
            "headers": {
              "Umb-Notifications": {
                "description": "The list of notifications produced during the request.",
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationHeaderModel"
                  },
                  "nullable": true
                }
              }
            }
          }
        },
        "security": [
          {
            "Backoffice-User": [ ]
          }
        ]
      }
    },
    "/umbraco/pagespeed-optimizer/v1/media-exception/by-media-key/{mediaKey}": {
      "get": {
        "tags": [
          "Page Speed Optimizer"
        ],
        "operationId": "GetByMediaKey",
        "parameters": [
          {
            "name": "mediaKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/MediaExceptionResponseModel"
                    }
                  ],
                  "nullable": true
                }
              }
            }
          },
          "401": {
            "description": "The resource is protected and requires an authentication token"
          },
          "403": {
            "description": "The authenticated user does not have access to this resource"
          }
        },
        "security": [
          {
            "Backoffice-User": [ ]
          }
        ]
      }
    },
    "/umbraco/pagespeed-optimizer/v1/media-exception/default-values": {
      "get": {
        "tags": [
          "Page Speed Optimizer"
        ],
        "operationId": "GetDefaultValues",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/DefaultValuesResponseModel"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "The resource is protected and requires an authentication token"
          },
          "403": {
            "description": "The authenticated user does not have access to this resource"
          }
        },
        "security": [
          {
            "Backoffice-User": [ ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "CreateMediaExceptionRequestModel": {
        "required": [
          "forceWebp",
          "mediaKey",
          "quality"
        ],
        "type": "object",
        "properties": {
          "mediaKey": {
            "type": "string",
            "format": "uuid"
          },
          "quality": {
            "type": "integer",
            "format": "int32"
          },
          "forceWebp": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "DefaultValuesResponseModel": {
        "required": [
          "defaultImageQuality",
          "enabled",
          "forceWebP"
        ],
        "type": "object",
        "properties": {
          "defaultImageQuality": {
            "type": "integer",
            "format": "int32"
          },
          "enabled": {
            "type": "boolean"
          },
          "forceWebP": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "EventMessageTypeModel": {
        "enum": [
          "Default",
          "Info",
          "Error",
          "Success",
          "Warning"
        ],
        "type": "string"
      },
      "MediaExceptionResponseModel": {
        "required": [
          "forceWebp",
          "id",
          "mediaKey",
          "quality"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "mediaKey": {
            "type": "string",
            "format": "uuid"
          },
          "quality": {
            "type": "integer",
            "format": "int32"
          },
          "forceWebp": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "NotificationHeaderModel": {
        "required": [
          "category",
          "message",
          "type"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/EventMessageTypeModel"
          }
        },
        "additionalProperties": false
      },
      "UpdateMediaExceptionRequestModel": {
        "required": [
          "forceWebp",
          "quality"
        ],
        "type": "object",
        "properties": {
          "quality": {
            "type": "integer",
            "format": "int32"
          },
          "forceWebp": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Backoffice-User": {
        "type": "oauth2",
        "description": "Umbraco Authentication",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "/umbraco/management/api/v1/security/back-office/authorize",
            "tokenUrl": "/umbraco/management/api/v1/security/back-office/token",
            "scopes": { }
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Page Speed Optimizer"
    }
  ]
}